+++ libsolutil/Exceptions.h
@@ -18,6 +18,8 @@
 
 #pragma once
 
+#include <liblangutil/SourceLocation.h>
+
 #include <boost/exception/exception.hpp>
 #include <boost/exception/info.hpp>
 #include <boost/exception/info_tuple.hpp>
@@ -39,6 +41,9 @@ struct Exception: virtual std::exception, virtual boost::exception
 
 	/// @returns the errinfo_comment of this exception.
 	std::string const* comment() const noexcept;
+
+	/// @returns the errinfo_sourceLocation of this exception
+	langutil::SourceLocation sourceLocation() const noexcept;
 };
 
 /// Throws an exception with a given description and extra information about the location the