+++ scripts/error_codes.py
@@ -201,7 +201,7 @@ def examine_id_coverage(top_dir, source_id_to_file_names, new_ids_only=False):
                 # The warning may or may not exist in a compiler build.
         "4591", # "There are more than 256 warnings. Ignoring the rest."
                 # Due to 3805, the warning lists look different for different compiler builds.
-        "1834", # Unimplemented feature error, as we do not test it anymore via cmdLineTests
+        "1920", # Unimplemented feature error from YulStack (currently there are no UnimplementedFeatureErrors thrown by libyul)
         "1180", # SMTChecker, covered by CL tests
         "2339", # SMTChecker, covered by CL tests
         "2961", # SMTChecker, covered by CL tests