+++ test/libsolidity/syntaxTests/array/nested_calldata_storage2.sol
@@ -6,4 +6,4 @@ contract C {
 }
 
 // ----
-// UnimplementedFeatureError: Copying nested calldata dynamic arrays to storage is not implemented in the old code generator.
+// UnimplementedFeatureError 1834: Copying nested calldata dynamic arrays to storage is not implemented in the old code generator.