+++ test/libsolidity/syntaxTests/experimental/builtin/builtin_type_definition.sol
@@ -33,7 +33,6 @@ contract C {
 // EVMVersion: >=constantinople
 // compileViaYul: true
 // ----
-// UnimplementedFeatureError: No support for calling functions pointers yet.
 // Warning 2264: (0-29): Experimental features are turned on. Do not use experimental features on live deployments.
 // Info 4164: (31-61): Inferred type: void
 // Info 4164: (63-93): Inferred type: bool
@@ -90,3 +89,4 @@ contract C {
 // Info 4164: (563-574): Inferred type: (bool, word) -> word
 // Info 4164: (563-567): Inferred type: (?bm:type, ?bn:type)
 // Info 4164: (575-576): Inferred type: (bool, word)
+// UnimplementedFeatureError 1834: No support for calling functions pointers yet.