+++ test/libsolidity/syntaxTests/types/rational_number_literal_to_fixed_implicit.sol
@@ -13,4 +13,4 @@ contract C {
     }
 }
 // ----
-// UnimplementedFeatureError: Not yet implemented - FixedPointType.
+// UnimplementedFeatureError 1834: Not yet implemented - FixedPointType.